gcp/run_contact_pipeline/cloud_run_functions/requirements.txt
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.0
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
deprecation==2.1.0
Flask==3.1.0
functions-framework==3.8.2
google-api-core==2.24.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-cloud-secret-manager==2.22.1
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
gspread==6.1.4
gspread-dataframe==4.0.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgpack==1.1.0
numpy==2.2.2
oauthlib==3.2.2
packaging==24.1
pandas==2.2.3
pendulum==3.0.0
prompt_toolkit==3.0.47
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
python-box==7.3.2
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
radian==0.6.13
rchitect==0.4.7
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
six==1.16.0
time-machine==2.16.0
toml==0.10.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
